+++ src/client/src/pages/Register/index.ts
